requirements.txt
streamlit==1.28.0
gspread==5.11.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
pandas==2.0.3
numpy==1.24.3